feat: refactored penalty and fitness functions to Penalty.java
