+++ src/main/java/com/course_scheduling/app/App.java
@@ -14,33 +14,33 @@ import java.time.LocalDateTime;
  *
  */
 public class App {
-    
+
     private final FileManager fileManager = new FileManager();
 
     public static void main(String[] args) {
-        
+
         App app = new App();
-        // run GA algorithm
-//        Scheduler driver = new Scheduler();
-//        driver.runAlgorithm();
+//         run GA algorithm
+        Scheduler driver = new Scheduler();
+        driver.runAlgorithm();
 
         //run PSO algorithm
-        app.runPsoAlgorithm("default", "limitFV", 100, 30);
+//        app.runPsoAlgorithm("default", "limitFV", 100, 30);
     }
-    
+
     private void runPsoAlgorithm(String datasetSize, String psoMode, int psoLimit, int psoParticles) {
         String courseDataset = "src/main/resources/dataset/processed/courses.csv";
-        switch(datasetSize) {
-            case "small" :
+        switch (datasetSize) {
+            case "small":
                 courseDataset = "src/main/resources/dataset/processed/courses_small.csv";
                 break;
-            case "medium" :
+            case "medium":
                 courseDataset = "src/main/resources/dataset/processed/courses_medium.csv";
                 break;
-            case "large" :
+            case "large":
                 courseDataset = "src/main/resources/dataset/processed/courses_large.csv";
                 break;
-            default :
+            default:
                 courseDataset = "src/main/resources/dataset/processed/courses.csv";
         }
         float[] coursesWeeklyHours;
@@ -55,7 +55,8 @@ public class App {
         possibleTimesDurations = pso.psoGetPossibleTimesDurations("src/main/resources/dataset/processed/timeslots.csv");
         possibleInstructors = pso.psoGetPossibleInstructors("src/main/resources/dataset/processed/instructors.csv");
         coursesPreferredTimes = pso.psoGetCoursesPreferredTimes("src/main/resources/dataset/processed/instructors.csv");
-        String logOutput = "";        
+        String logOutput = "";
+
         //initialize PSO; automatic stop after 15 minutes
         //"limitFV", 700, 30 => operation mode limit FV, 700 max penalty score, 30 particles
         //"limitIter", 1000, 30 => operation mode limit Iter, 1000 max iteration, 30 particles
@@ -69,7 +70,7 @@ public class App {
         DateTimeFormatter psoDTF = DateTimeFormatter.ofPattern("_yyyy_MM_dd_HH_mm_ss");
         LocalDateTime psoNow = LocalDateTime.now();
         String psoTimestamp = psoDTF.format(psoNow) + ".txt";
-        fileManager.createTextFile(logOutput, "Schedules-PSO"+psoTimestamp, "results/pso/");
+        fileManager.createTextFile(logOutput, "Schedules-PSO" + psoTimestamp, "results/pso/");
     }
 
 }
